added ; geoWithin geoIntersects nearSphere to SelectorBuilder
