+++ lib/mongo_dart_query.dart
@@ -2,5 +2,6 @@ library mongo_dart_query;
 
 import 'dart:convert' show json;
 import 'package:bson/bson.dart';
+import 'package:mongo_dart_query/src/geometry_obj.dart';
 part 'src/selector_builder.dart';
 part 'src/modifier_builder.dart';
